+++ tests/browser/test-ports.js
@@ -1,0 +1,4 @@
+module.exports = {
+  PARENT_PORT: 3001,
+  CHILD_PORT: 3002
+};